+++ src/core/scripting/lib-programming/@types/systemjs/extras/named-register.d.ts
