+++ src/core/scripting/lib-programming/executor/index.ts
